requirements.txt
pypiwin32
pycryptodome
hmac
pyasn1
